refactor(adapter-node-http): Move util methods to utils
