+++ packages/@pollyjs/adapter-node-http/tests/utils/get-buffer-from-stream.js
@@ -1,0 +1,13 @@
+export default function getBufferFromStream(stream) {
+  return new Promise(resolve => {
+    const chunks = [];
+
+    stream.on('data', chunk => {
+      chunks.push(chunk);
+    });
+
+    stream.on('end', () => {
+      resolve(Buffer.concat(chunks));
+    });
+  });
+}
