+++ packages/@pollyjs/adapter-node-http/tests/utils/get-response-from-request.js
@@ -1,0 +1,9 @@
+export default function getResponseFromRequest(req) {
+  return new Promise(resolve => {
+    req.on('response', res => {
+      resolve(res);
+    });
+
+    req.end();
+  });
+}
